requirements.txt
selenium==4.9.1
python-telegram-bot==20.3
python-dotenv==1.0.0
pytz==2023.3
pillow==10.0.0
